backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.1
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.2
openai>=1.3.0
chromadb>=0.4.15
numpy>=1.24.3
python-dotenv>=1.0.0
redis>=5.0.1
celery>=5.3.4
opencv-python>=4.8.1.78
aiofiles>=23.2.1
pydantic-settings>=2.1.0